Refactor duplicated logic across attributes
